Travis: For arm64 add apt packages and wheels from piwheels
This manually install the PyGame and numpy wheels as the architecture
detected by Python is aarch64 instea of armv6 or armv7.

Apt packages added from the Raspberry Pi instructions.

As piwheels doesn't yet support Requires-Python, some pacakges
would be installed with the latest version incompatible with
Python 3.5, so we cannot add the index yet.